Patched to remove some compiler warnings
git-svn-id: https://svn.code.sf.net/p/freemat/code/trunk@1854 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libGraphics/HandleImage.cpp
@@ -185,9 +185,10 @@ namespace FreeMat {
     for (int i=0;i<rows;i++) {
       QRgb *ibits = (QRgb*) img.scanLine(i);
       for (int j=0;j<cols;j++)
- 	ibits[j] = qRgba(255*dp[(i+j*rows)],255*dp[(i+j*rows)+rows*cols],
- 			 255*dp[(i+j*rows)+2*rows*cols],
-			 255*alpha[i+j*rows]);
+ 	ibits[j] = qRgba((int)(255*dp[(i+j*rows)]),
+			 (int)(255*dp[(i+j*rows)+rows*cols]),
+ 			 (int)(255*dp[(i+j*rows)+2*rows*cols]),
+			 (int)(255*alpha[i+j*rows]));
     }
   }
